Fruits, leaf scaling
had a hard time getting colliders to work properly. I tried attaching the colliders as children of the tree directly, but I wasn't able to get the positions to actually be relative to the tree. otherwise it works as far as raycasting goes, but i wasn't able to get both correct positioning and raycasting working, not sure what was different as far as the coordinate space for the rigidbodies